feat: comment smart-contracts functions.

+++ Backend/hardhat/contracts/Token.sol
@@ -3,32 +3,36 @@ pragma solidity >=0.7.0 <0.9.0;
 
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
+// Contract that creates the properties and the tokens for each property
 contract Token is ERC20 {
-    uint public amount;
+    uint256 public amount;
     uint256 public value;
     address public owner;
-    uint public tokenSupply;
+    uint256 public tokenSupply;
     address[] public tokenOwners;
 
-    mapping(address => mapping(address => uint)) internal waitingTokens;
+    mapping(address => mapping(address => uint256)) internal waitingTokens;
 
-    mapping(address => uint) internal tokenValue;
+    mapping(address => uint256) internal tokenValue;
 
+    // Starts the contract and sets the owner as the one who deployed it
     constructor(
         string memory _name,
         string memory _symbol,
         address _address,
-        uint _amount
+        uint256 _amount
     ) ERC20(_name, _symbol) {
         owner = _address;
         amount = _amount;
     }
 
+    // Checks if who is triggering the contract is the owner
     modifier isOwner() {
         require(msg.sender == owner);
         _;
     }
 
+    // This function mint tokens
     function claimTokens() public payable {
         require(
             tokenSupply < amount,
@@ -44,10 +48,11 @@ contract Token is ERC20 {
         tokenSupply += toReceive;
     }
 
+    //
     function approveWithValue(
         address _spender,
         uint256 _amount,
-        uint _value
+        uint256 _value
     ) public returns (bool) {
         approve(_spender, _amount);
         tokenValue[msg.sender] = _value;
@@ -55,6 +60,7 @@ contract Token is ERC20 {
         return true;
     }
 
+    // This function buy tokens
     function buyToken(address _owner, uint256 _amount) public payable {
         require(allowance(_owner, owner) > 0, "Not on sale");
         require(msg.value >= tokenValue[_owner]);
@@ -62,6 +68,7 @@ contract Token is ERC20 {
         waitingTokens[msg.sender][_owner];
     }
 
+    // This function allows the owner to sell tokens
     function transferTokenTo(address _owner, address _receiver) public isOwner {
         require(
             waitingTokens[_receiver][_owner] > 0,
@@ -79,6 +86,7 @@ contract Token is ERC20 {
         return tokenOwners;
     }
 
+    // This function burns all the tokens
     function burnTokens() public isOwner returns (bool) {
         require(
             balanceOf(_msgSender()) == tokenSupply,
@@ -88,6 +96,7 @@ contract Token is ERC20 {
         return true;
     }
 
+    // This function sets the value of the token
     function _afterTokenTransfer(
         address from,
         address to,
